% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/screenshot.R
\name{screenshot}
\alias{screenshot}
\title{Embed a screenshot.}
\usage{
screenshot(path, dpi = NULL)
}
\arguments{
\item{path}{Path to screenshot.}

\item{dpi}{DPI of image. You can leave this blank if the dpi is stored
in the png metadata}
}
\description{
This embeds a png screenshot into the document. Generally you should use
\code{echo = FALSE} on such blocks. By convention, screenshots should
live in a \code{screenshots/} directory.
}

